Sort user and usergroup by most match order
  fixes #17859
